# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: client manager
  description: do some CRUD
  contact:
    name: peiwen.wang
    email: pw5508@mun.ca

host: localhost:5000
basePath: /v1
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /ping:
    get:
      x-swagger-router-controller: ping
      description:
        health check
      operationId: ping
      responses:
        "200":
          description: ping response
          schema:
            type: string
            items:
              $ref: '#/definitions/PingResponse'


  /clients:
    get:
      description: |
        Returns clients from the system that meet certain requirement
      x-swagger-router-controller: client
      operationId: getClient
      parameters:
        - name: name
          in: query
          description: name to filter by
          required: false
          type: string
        - name: phone
          in: query
          description: phone to filter by
          required: false
          type: string
        - name: email
          in: query
          description: email to filter by
          required: false
          type: string
        - name: minage
          in: query
          description: describe age range to filter by
          required: false
          type: string
        - name: maxage
          in: query
          description: describe age range to filter by
          required: false
          type: string    
        - name: limit
          in: query
          description: maximum number of results to return
          required: false
          type: integer
          format: int32
      responses:
        "200":
          description: clents response
          schema:
            type: array
            items:
              $ref: '#/definitions/Client'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'

    post:
      x-swagger-router-controller: client
      description: Creates a new client
      operationId: createClient

      parameters:
        - in: body
          name: form
          description: a form of client info
          schema:
            $ref: '#/definitions/NewClient'

      responses:
        "200":
          description: success response
          schema:
            $ref: '#/definitions/ClientCreatedResponse'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
  /clients/{id}:
    put:
      description: update a client
      x-swagger-router-controller: client
      operationId: updateClient
      parameters:
        - name: id
          in: path
          description: ID of client to update
          required: true
          type: string
        - name: form
          in: body
          description: a form of client info
          schema:
            $ref: '#/definitions/NewClient'
      responses:
        "204":
          description:  success response
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      description: deletes a single clien based on the ID supplied
      x-swagger-router-controller: client
      operationId: deleteClient
      parameters:
        - name: id
          in: path
          description: ID of clien to delete
          required: true
          type: string
      responses:
        "204":
          description: client deleted
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'

  /accounts:
        get:
          description:
            Returns accounts from the system that meet certain requirement
          x-swagger-router-controller: account
          operationId: getAccount
          parameters:
            - name: number
              in: query
              description: number to filter by
              required: false
              type: string
            - name: type
              in: query
              description: type to filter by
              required: false
              type: string
            - name: status
              in: query
              description: status to filter by
              required: false
              type: string
            - name: limit
              in: query
              description: maximum number of results to return
              required: false
              type: integer
              format: int32
            - name: force
              in: query
              description: if force is set to true it means do not get cache
              required: false
              type: string
          responses:
            "200":
              description: account response
              schema:
                type: array
                items:
                  $ref: '#/definitions/Account'
            default:
              description: unexpected error
              schema:
                $ref: '#/definitions/Error'

        post:
              x-swagger-router-controller: account
              description: Creates a new account
              operationId: createAccount

              parameters:
                - in: body
                  name: form
                  description: a form of client info
                  schema:
                    $ref: '#/definitions/NewAccount'

              responses:
                "200":
                  description: success response
                  schema:
                    $ref: '#/definitions/AccountCreatedResponse'
                default:
                  description: unexpected error
                  schema:
                    $ref: '#/definitions/Error'

  /accounts/{id}:
        put:
          description: update an account
          x-swagger-router-controller: account
          operationId: updateAccount
          parameters:
            - name: id
              in: path
              description: number of the account to update
              required: true
              type: string
            - name: form
              in: body
              description: a form of account info
              schema:
                $ref: '#/definitions/NewAccount'
          responses:
            "204":
              description:  success response
            default:
              description: unexpected error
              schema:
                $ref: '#/definitions/Error'
        delete:
          description: deletes an account based on the account number supplied
          x-swagger-router-controller: account
          operationId: deleteAccount
          parameters:
            - name: id
              in: path
              description: account number of the account to delete
              required: true
              type: string
          responses:
            "204":
              description: account deleted
            default:
              description: unexpected error
              schema:
                $ref: '#/definitions/Error'

definitions:
  Client:
    required:
      - name
    properties:
      name:
        type: string
      address:
        type: string
      postalCode:
        type: string
      phone:
        type: string
      email:
        type: string
      dob:
        type: string  
  NewClient:
    type: object
    required:
      - name
    properties:
      name:
        type: string
      address:
        type: string
      postalCode:
        type: string
      phone:
        type: string
      email:
        type: string
      dob:
        type: string

  NewAccount:
      type: object
      required:
        - type
        - status
      properties:
        type:
          type: string
        status:
          type: string

  Account:
        type: object
        required:
          - number
        properties:
          number:
            type: string
          type:
            type: string
          status:
            type: string

  ClientCreatedResponse:
    properties:
        id:
          type: string

  AccountCreatedResponse:
      properties:
          number:
            type: string
  
  Error:
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string

  PingResponse:
    properties:
        version:
          type: string